+++ rules/react.js
@@ -71,8 +71,6 @@ module.exports = {
 		'react/jsx-pascal-case': 2,
 		// Enforce props alphabetical sorting
 		'react/jsx-sort-props': 0,
-		// Validate spacing before closing bracket in JSX
-		'react/jsx-space-before-closing': [2, 'always'],
 		// Validate whitespace in and around the JSX opening and closing brackets
 		'react/jsx-tag-spacing': [2, {
 			'closingSlash': 'never',